Only one edge case remains!
This change introduces a concept of wildcards.
Wildcards are the cards which are not a part of your category solution.
They are what one should compare for the run-off if two hands are
matched in both category and score.

Only remaining edge case is to handle the two_pair comparison.